Add files via upload
In this code, we use the 'Servo' library to control the servo motor. We connect a potentiometer to pin A0 (analog input pin) and the servo motor to pin 9 (you can choose another available pin). We read the value from the potentiometer and map it to an angle from 0 to 180 degrees using the 'map()' function. Then we pass this angle to the servo motor using the 'write()' method.